utils.git: Add and test resolve_symbolic_ref() and get_current_branch()
To test that resolve_symbolic_ref() works correctly, created repository
should not be in detached HEAD state, but the test for checkout_revision()
did just that by accident, which was then fixed.  The fix uses newly
added get_current_branch(), but it could use also newly introduced
self.default_branch - though it would be less forgiving solution.

To ensure that we know what the result of get_current_branch() should
be, we need to set the name of default branch, because in future
versions of Git the name of default branch may change from 'master' to
'main'.

TODO: add docstrings for get_current_branch() and resolve_symbolic_ref().