utils.git: Add to_oid(), reimplement is_valid_commit() using it
The to_oid() method will be useful in test to check that returned SHA-1
matches what is expected, even if repository creation does not ensure
stable SHA-1.  Instead of comparing SHA-1s, with to_oid() we can check
that known commit (like "v1", or "HEAD", or "HEAD^") resolves to the
same SHA-1 as the one we got.

While at it, remove accidentally added line between methods.