dvc.lock: Run "clone_repos" stage
*Note* that while running this stage, `git clone` asked 8 times about
credentials.  This might mean that 8 of repositories from DevGPT dataset
are currently not public on GitHub.

While at it, update command in "download_DevGPT" to match what it is
now in dvc.yaml (the stage was run with pre-commit version of it).

The data/repositories_download_status.json, which is not tracked by Git,
got marked as ignored in data/.gitignore.