compute_changes_survival.py: Add timestamps to line survival data
The main process_commits() function was modified to gather commit info
(per project) that it got as a side effect of running the reverse blame,
and pass it to process_commit_changed_lines().

The blamed_commits_data field was added to GptCommitInfo named tuple,
and process_single_commit() was modified to return this data that it
gathers.  While at it, GptCommitInfo was modified to use default value
of None for optional fields, which allows simpler initialization where
only partial data is available.

The most important change was modifying process_commit_changed_lines()
so that it adds author timestamp and committer timestamp for all commits
this data mentions / includes.