compute_changes_survival.py: Update and improve handling missing things
Update handling of the case where repository was not cloned to make use
of reponame_to_repo_path().  This updates the code to make use of new
cloned repositories layout, and to match changes to
load_repositories_json().

Add handling of the case when 'Sha' column is N/A = None, which can
be the case for pr_sharings_to_agg.py results.