pr_sharings_to_agg.py: Add another output file
Before this change this script was computing and saving aggregated data
from DevGPT dataset for each pull request in most recent pr sharings
(to 'pr_sharings_df.csv').  The 'Sha' column in this dataset was
SHA-1 identifier of the merge commit (if it exists), retrieved via
GitHub API.  It later turned out that even after limiting storing
the retrieved value in the 'Sha' to the cases where GitHub API confirmed
that pull request was merged in, the 'Sha' commit might be missing from
the cloned project (possibly cloned earlier).

Now the script also saves the very similar dataframe, but with 'Sha'
column set to every commit in 'CommitSha' list (of commits that make up
the pull request).  This split dataframe is saved in the
'pr_sharings_split_commit_df.csv' file.