issue_sharings_to_agg.py: Find closing commit for issues that have one
Retrieving closing pull request (and its merge commit) or closing commit
is done by querying GitHub using the GraphQL API v4 using timelineItems
and filtering on event with state CLOSED_EVENT, see
https://stackoverflow.com/a/62294269/46058

Note that this information is not available from GitHub REST API; the
result of PyGithub's get_events() method does not include the
information needed.

The alternative would be to search through all pull requests (with
PyGithub) and through all commits (for example with `git log --grep`)
for mentions of the issue (for example in the "\b#<number>\b" form).
The search could be done with a crafter regexp, or with regexp to detect
mentioning issue and checking if retrieved issue number match, see e.g.:
https://stackoverflow.com/questions/42742810/speed-up-millions-of-regex-replacements-in-python-3
https://www.rexegg.com/regex-optimizations.html