compute_changes_survival.py: Fix issue with pathnames containing spaces
The 'previous' line in `git blame` or `git blame --reverse` output
consists of SHA-1 of commit and of file pathname, separated by a single
space.  The file pathname can include space, which the coded didn't
take into account, resulting in the following exception being thrown:

      prev_commit, prev_file = change_line_info['previous'].split(' ')
      ^^^^^^^^^^^^^^^^^^^^^^
  ValueError: too many values to unpack (expected 2)

The fix is to simply use `.split(' ', maxsplit=1)`.