utils.git: Implement optimization for added files in changes_survival()
It is not that much of optimization; the diff is still parsed, it is
just that for pure file additions `git blame --reverse` works on whole
files, being called without `-L <begin>,<end>` file range parameter(s).

This optimization is optional, turned on only when `addition_optim`
parameter is true.