test_utils_git.py: Add "changes survival from v2" subtest
This subtest is added to test_changes_survival() in GitTestCase.
It will be used to check correctness of handling renames; currently
changes in 'v2' include pure rename with no change to the file,
which means that this file is not present in changes_survival()
output - because there were no changes.