init.bash: Configure DVC to not need Git, if outside repo
DVC works best in a Git repository, and is by default configured to
require it.  If this project was downloaded instead of being cloned,
we might be not inside any Git repository.  In that case, configure DVC
(locally) to not require to be run from inside git repo.

NOTE: not tested.